[frontend] More Documentation
