+++ frontend/src/app/layout.tsx
@@ -1,8 +1,15 @@
-export default function RootLayout({
-  children,
-}: {
+interface Props {
   children: React.ReactNode
-}) {
+}
+
+/**
+ * @name RootLayout
+ * @description The root layout component, top level component for all pages
+ *
+ * @param {React.ReactNode} children - The children to render
+ * @returns {JSX.Element} The root layout component
+ */
+export default function RootLayout({ children }: Props): React.ReactElement {
   return (
     <html suppressHydrationWarning>
       <head />
