+++ frontend/src/app/robots.ts
@@ -1,5 +1,11 @@
 import { MetadataRoute } from 'next';
 
+/**
+ * @name robots
+ * @description Generate the robots.txt file for the site
+ * 
+ * @returns {MetadataRoute.Robots} The robots.txt configuration 
+ */
 export default function robots(): MetadataRoute.Robots {
   return {
     rules: [
